Replace discretizedswaption by discretizedswaption2
