+++ ql/pricingengines/swaption/treeswaptionengine.hpp
@@ -3,7 +3,6 @@
 /*
  Copyright (C) 2001, 2002, 2003 Sadruddin Rejeb
  Copyright (C) 2005, 2007 StatPro Italia srl
- Copyright (C) 2022 Ralf Konrad Eckel
 
  This file is part of QuantLib, a free-software/open-source library
  for financial quantitative analysts and developers - http://quantlib.org/
@@ -27,12 +26,7 @@
 #define quantlib_tree_swaption_engine_hpp
 
 #include <ql/instruments/swaption.hpp>
-#include <ql/math/functional.hpp>
 #include <ql/pricingengines/latticeshortratemodelengine.hpp>
-#include <ql/pricingengines/swaption/discretizedswaption.hpp>
-#include <ql/pricingengines/swaption/discretizedswaption2.hpp>
-#include <utility>
-
 
 namespace QuantLib {
 
@@ -47,8 +41,7 @@ namespace QuantLib {
 
         \test calculations are checked against cached results
     */
-    template <class TDiscretizedSwaption>
-    class GenericTreeSwaptionEngine
+    class TreeSwaptionEngine
     : public LatticeShortRateModelEngine<Swaption::arguments,
                                          Swaption::results> {
       public:
@@ -57,18 +50,15 @@ namespace QuantLib {
                   model cannot provide one itself.
         */
         //@{
-        GenericTreeSwaptionEngine(
-            const ext::shared_ptr<ShortRateModel>&,
-            Size timeSteps,
-            Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
-        GenericTreeSwaptionEngine(
-            const ext::shared_ptr<ShortRateModel>&,
-            const TimeGrid& timeGrid,
-            Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
-        GenericTreeSwaptionEngine(
-            const Handle<ShortRateModel>&,
-            Size timeSteps,
-            Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
+        TreeSwaptionEngine(const ext::shared_ptr<ShortRateModel>&,
+                           Size timeSteps,
+                           Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
+        TreeSwaptionEngine(const ext::shared_ptr<ShortRateModel>&,
+                           const TimeGrid& timeGrid,
+                           Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
+        TreeSwaptionEngine(const Handle<ShortRateModel>&,
+                           Size timeSteps,
+                           Handle<YieldTermStructure> termStructure = Handle<YieldTermStructure>());
         //@}
         void calculate() const override;
 
@@ -76,83 +66,6 @@ namespace QuantLib {
         Handle<YieldTermStructure> termStructure_;
     };
 
-    using TreeSwaptionEngine = GenericTreeSwaptionEngine<DiscretizedSwaption>;
-    using TreeSwaptionEngine2 = GenericTreeSwaptionEngine<DiscretizedSwaption2>;
-
-    template <class TDiscretizedSwaption>
-    GenericTreeSwaptionEngine<TDiscretizedSwaption>::GenericTreeSwaptionEngine(
-        const ext::shared_ptr<ShortRateModel>& model,
-        Size timeSteps,
-        Handle<YieldTermStructure> termStructure)
-    : LatticeShortRateModelEngine<Swaption::arguments, Swaption::results>(model, timeSteps),
-      termStructure_(std::move(termStructure)) {
-        registerWith(termStructure_);
-    }
-
-    template <class TDiscretizedSwaption>
-    GenericTreeSwaptionEngine<TDiscretizedSwaption>::GenericTreeSwaptionEngine(
-        const ext::shared_ptr<ShortRateModel>& model,
-        const TimeGrid& timeGrid,
-        Handle<YieldTermStructure> termStructure)
-    : LatticeShortRateModelEngine<Swaption::arguments, Swaption::results>(model, timeGrid),
-      termStructure_(std::move(termStructure)) {
-        registerWith(termStructure_);
-    }
-
-    template <class TDiscretizedSwaption>
-    GenericTreeSwaptionEngine<TDiscretizedSwaption>::GenericTreeSwaptionEngine(
-        const Handle<ShortRateModel>& model,
-        Size timeSteps,
-        Handle<YieldTermStructure> termStructure)
-    : LatticeShortRateModelEngine<Swaption::arguments, Swaption::results>(model, timeSteps),
-      termStructure_(std::move(termStructure)) {
-        registerWith(termStructure_);
-    }
-
-    template <class TDiscretizedSwaption>
-    void GenericTreeSwaptionEngine<TDiscretizedSwaption>::calculate() const {
-
-        QL_REQUIRE(arguments_.settlementMethod != Settlement::ParYieldCurve,
-                   "cash settled (ParYieldCurve) swaptions not priced with "
-                   "TreeSwaptionEngine");
-        QL_REQUIRE(!model_.empty(), "no model specified");
-
-        Date referenceDate;
-        DayCounter dayCounter;
-
-        ext::shared_ptr<TermStructureConsistentModel> tsmodel =
-            ext::dynamic_pointer_cast<TermStructureConsistentModel>(*model_);
-        if (tsmodel != nullptr) {
-            referenceDate = tsmodel->termStructure()->referenceDate();
-            dayCounter = tsmodel->termStructure()->dayCounter();
-        } else {
-            referenceDate = termStructure_->referenceDate();
-            dayCounter = termStructure_->dayCounter();
-        }
-
-        TDiscretizedSwaption swaption(arguments_, referenceDate, dayCounter);
-        ext::shared_ptr<Lattice> lattice;
-
-        if (lattice_ != nullptr) {
-            lattice = lattice_;
-        } else {
-            std::vector<Time> times = swaption.mandatoryTimes();
-            TimeGrid timeGrid(times.begin(), times.end(), timeSteps_);
-            lattice = model_->tree(timeGrid);
-        }
-
-        std::vector<Time> stoppingTimes(arguments_.exercise->dates().size());
-        for (Size i = 0; i < stoppingTimes.size(); ++i)
-            stoppingTimes[i] = dayCounter.yearFraction(referenceDate, arguments_.exercise->date(i));
-
-        swaption.initialize(lattice, stoppingTimes.back());
-
-        Time nextExercise = *std::find_if(stoppingTimes.begin(), stoppingTimes.end(),
-                                          greater_or_equal_to<Time>(0.0));
-        swaption.rollback(nextExercise);
-
-        results_.value = swaption.presentValue();
-    }
 }
